Creating logical device

+++ source/monomyst/vulkan/queues.d
@@ -4,7 +4,7 @@ import std.typecons;
 
 struct QueueFamilyIndices
 {
-    Nullable!ulong graphicsFamily;
+    Nullable!uint graphicsFamily;
 
     bool isComplete ()
     {
